requirements.txt
Flask>=3.0.0
numpy>=1.26.0
scipy>=1.11.0
plotly>=5.18.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
